Move TransformerGuard to namespace transformation
